:scroll: Add SVG of FOSS minor promo poster (closes FOSSRIT/tasks#83)
This commit adds the mostly-converted "A Year in Open Source" poster SVG
used to advertise FOSS courses for the coming year at RIT. The
information in here is dated and no longer relevant, but it is the only
"open source" version of this poster available. It is slightly improved
from what was shared in FOSSRIT/tasks#83 because I ported most of the
proprietary fonts over to open source equivalents that can be installed
on any operating system without paying a license fee.

Since the academic year is almost over and the issue is over a year old,
I am closing out this issue from this commit and we can reopen when we
are ready to plan out a new poster design.

Closes FOSSRIT/tasks#83.

Signed-off-by: Justin W. Flory <git@jwf.io>